#656: Perform refresh in Eclipse after M2E execution of jaxws-maven-plugin

+++ jaxws-ri/extras/jaxws-maven-plugin/src/main/java/org/jvnet/jax_ws_commons/jaxws/AbstractWsGenMojo.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2012, 2022 Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2012, 2023 Oracle and/or its affiliates. All rights reserved.
  * Copyright 2006 Guillaume Nodet
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -139,15 +139,18 @@ abstract class AbstractWsGenMojo extends AbstractJaxwsMojo {
     }
 
     protected void processSei(String sei) throws MojoExecutionException {
-        getLog().info("Processing: " + sei);
-        List<String> args = getWsGenArgs(sei, true);
-        getLog().info("jaxws:wsgen args: " + args);
-        exec(args);
-        if (metadata != null) {
-            try {
-                FileUtils.copyFileToDirectory(metadata, getClassesDir());
-            } catch (IOException ioe) {
-                throw new MojoExecutionException(ioe.getMessage(), ioe);
+        if (buildContext.hasDelta(new File(getClassesDir(), sei.replace('.', '/') + ".class"))
+                || (metadata != null && buildContext.hasDelta(metadata))) {
+            getLog().info("Processing: " + sei);
+            List<String> args = getWsGenArgs(sei, true);
+            getLog().info("jaxws:wsgen args: " + args);
+            exec(args);
+            if (metadata != null) {
+                try {
+                    FileUtils.copyFileToDirectory(metadata, getClassesDir());
+                } catch (IOException ioe) {
+                    throw new MojoExecutionException(ioe.getMessage(), ioe);
+                }
             }
         }
     }
